removed constants support entirely
during implementation of object (record) support for constants, it turned out that
object constants cannot be supported in objc interfaces.
The current Djinni implementation seems to just hide this problem
by not generating the required code and not giving an error.

Given the fact that also the implementation overhead for supporting constants
is quite big compared to the little edge use-cases
that it might be useful in, I decided to dump the feature all together. This also
saves me from worrying questions, for example on how to implement list and dict
support for constants, or how to support enum, flags and records support for
external types.

All in all the feature tends to blur the lines between the IDL being just a
interface definition language to it starting to become a programming language.
IMO PyDjinni should not continue this approach and thereby the feature was removed.